+++ chromium/testing/test_env.py
@@ -86,6 +86,9 @@ def get_sanitizer_env(cmd, asan, lsan, msan, tsan, cfi_diag):
   else:
     symbolization_options = []
 
+  # Leverage sanitizer to print stack trace on abort (e.g. assertion failure).
+  symbolization_options.append('handle_abort=1')
+
   if asan:
     asan_options = symbolization_options[:]
     if lsan:
@@ -216,6 +219,24 @@ def run_command(argv, env=None, cwd=None, log=True):
   return wait_with_signals(process)
 
 
+def run_command_output_to_handle(argv, file_handle, env=None, cwd=None):
+  """Run command and stream its stdout/stderr both to |file_handle|.
+
+  Also forward_signals to obey
+  https://chromium.googlesource.com/infra/luci/luci-py/+/master/appengine/swarming/doc/Bot.md#graceful-termination_aka-the-sigterm-and-sigkill-dance
+
+  Returns:
+    integer returncode of the subprocess.
+  """
+  print('Running %r in %r (env: %r)' % (argv, cwd, env))
+  process = subprocess.Popen(
+      argv, env=env, cwd=cwd, stderr=file_handle, stdout=file_handle)
+  forward_signals([process])
+  exit_code = wait_with_signals(process)
+  print('Command returned exit code %d' % exit_code)
+  return exit_code
+
+
 def wait_with_signals(process):
   """A version of process.wait() that works cross-platform.
 
@@ -250,6 +271,8 @@ def forward_signals(procs):
   assert all(isinstance(p, subprocess.Popen) for p in procs)
   def _sig_handler(sig, _):
     for p in procs:
+      if p.poll() is not None:
+        continue
       # SIGBREAK is defined only for win32.
       if sys.platform == 'win32' and sig == signal.SIGBREAK:
         p.send_signal(signal.CTRL_BREAK_EVENT)
@@ -259,7 +282,7 @@ def forward_signals(procs):
     signal.signal(signal.SIGBREAK, _sig_handler)
   else:
     signal.signal(signal.SIGTERM, _sig_handler)
-
+    signal.signal(signal.SIGINT, _sig_handler)
 
 def run_executable(cmd, env, stdoutfile=None):
   """Runs an executable with:
@@ -311,10 +334,17 @@ def run_executable(cmd, env, stdoutfile=None):
   cmd[0] = cmd[0].replace('/', os.path.sep)
   cmd = fix_python_path(cmd)
 
+  # We also want to print the GTEST env vars that were set by the caller,
+  # because you need them to reproduce the task properly.
+  env_to_print = extra_env.copy()
+  for env_var_name in ('GTEST_SHARD_INDEX', 'GTEST_TOTAL_SHARDS'):
+      if env_var_name in env:
+          env_to_print[env_var_name] = env[env_var_name]
+
   print('Additional test environment:\n%s\n'
         'Command: %s\n' % (
         '\n'.join('    %s=%s' %
-            (k, v) for k, v in sorted(extra_env.iteritems())),
+            (k, v) for k, v in sorted(env_to_print.iteritems())),
         ' '.join(cmd)))
   sys.stdout.flush()
   env.update(extra_env or {})
